backend/requirements.txt
# Flask and Web
Flask==3.1.2
flask-cors==6.0.1
Werkzeug==3.1.3

# Database
SQLAlchemy==2.0.44

# OpenAI for AI-powered chat
openai>=1.0.0

# Data processing
pandas==2.3.3
numpy==2.3.4
netCDF4==1.7.3
xarray==2025.10.1

# HTTP requests
requests>=2.31.0
certifi==2025.10.5

# Visualization (optional, for future use)
plotly==6.4.0

# Environment and utilities
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Time/Date handling
cftime==1.6.5

# Other dependencies
blinker==1.9.0
click==8.3.0
colorama==0.4.6
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
narwhals==2.10.2
packaging==25.0
six==1.17.0
typing_extensions==4.15.0
